listo modalAsk eliminar, solo tiene que ver la eliminacion reflejada y mostrar mensaje

+++ app/Http/Controllers/ProductController.php
@@ -177,7 +177,7 @@ class ProductController extends Controller
     {
         sleep(1);
         $product = Product::findOrFail($id);
-        if ($product->image_url !== '/image_url/default.jpeg') {
+        if ($product->image_url !== '/image_url/default.jpeg' && $product->image_url !== 'image_url/default.jpeg' ) {
             Storage::disk('public')->delete($product->image_url);
         }
         $product->delete();
